agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = ""
authors = [
    {name = "orca-copilotkit",email = "testapp@copilotkit.ai"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "openai (>=1.50.0,<2.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "fastapi (>=0.115.0,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "langgraph (>=1.0.0,<2.0.0)",
    "langchain-core (>=0.3.78,<0.4.0)",
    "langchain-openai (>=0.2.0,<0.3.0)",
    "copilotkit (==0.1.58)",
    "requests (>=2.31.0,<3.0.0)",
    "httpx (>=0.27.0,<0.28.0)"
]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
